Move data stream lifecycle related actions to server
